Add files via upload
# Bulb Animation

This code creates a simple animation of a light bulb turning on and off. 
The animation is created using two images, one of a light bulb that is off and one of a light bulb that is on. 
When the user clicks the "ON" button, the image of the light bulb that is on is displayed and the background 
colour of the page changes to yellow. 
When the user clicks the "OFF" button, the image of the light bulb that is off is 
displayed and the background color of the page changes to black.

## Step-by-Step Explanation

### HTML

The HTML code for this project is relatively simple. 
It consists of a `<div>` element that contains the light bulb image, 
two buttons, and a heading. The `<div>` element is centred horizontally and 
vertically using the `justify-content` and `align-items` properties. The' margin' property positions the buttons below the light bulb image. Using the ' margin-top ' property, the heading
is positioned above the light bulb image.